refactor(devtools): consume the shared jest presets and render fixtures
shell, feature-flags and transport-panel each carried their own copy of the jest
configs, setup files, React Native mocks and themed render helper. Three of those
files were byte-identical and the rest had drifted. They now call the
@support/jest-devtools-fixtures factories and keep a one-line re-export at
jest/render; shell keeps renderWithNavigation and renderScreen, which need its
own navigation and shell context.

Also fixes two latent gaps: transport-panel was missing the @gorhom/bottom-sheet
mock declaration and resolved its native render helper only through the
react-native preset's haste platform list, so its native tests now import
jest/render.native explicitly. babel.config.js stays in each package — babel
resolves its config from the jest rootDir, and babel-jest needs the React Native
preset to parse the Flow-typed files inside react-native itself.

The web tsconfigs now name @testing-library/jest-dom in types: the deleted
jest.setup.ts was what registered its matcher types.
