+++ devtools/feature-flags/jest.config.js
@@ -1,42 +1,1 @@
-module.exports = {
-  testEnvironment: "jsdom",
-  roots: ["<rootDir>/src"],
-  transform: {
-    "^.+\\.(t|j)sx?$": [
-      "@swc/jest",
-      {
-        jsc: {
-          target: "esnext",
-          transform: {
-            react: {
-              runtime: "automatic",
-            },
-          },
-        },
-      },
-    ],
-  },
-  moduleFileExtensions: [
-    "web.tsx",
-    "web.ts",
-    "tsx",
-    "ts",
-    "js",
-    "jsx",
-    "json",
-    "node",
-    "native.tsx",
-    "native.ts",
-  ],
-  testPathIgnorePatterns: ["\\.native\\.test\\."],
-  modulePaths: ["<rootDir>"],
-  setupFilesAfterEnv: ["<rootDir>/jest.setup.ts"],
-  transformIgnorePatterns: [
-    "node_modules/.pnpm/(?!(@ledgerhq\\+lumen-ui-react|@ledgerhq\\+lumen-design-core|@ledgerhq\\+lumen-utils-shared))",
-  ],
-  coverageReporters: ["json", ["lcov", { file: "lcov.info", projectRoot: "../../" }], "text"],
-  reporters: [
-    "default",
-    ["jest-sonar", { outputName: "sonar-executionTests-report.xml", reportedFilePath: "absolute" }],
-  ],
-};
+module.exports = require("@support/jest-devtools-fixtures").createWebJestConfig();
