+++ devtools/feature-flags/package.json
@@ -23,6 +23,7 @@
     "@shared/feature-flags": "workspace:^"
   },
   "devDependencies": {
+    "@support/jest-devtools-fixtures": "workspace:*",
     "@testing-library/react": "catalog:",
     "@testing-library/react-native": "catalog:",
     "@testing-library/dom": "catalog:",
