--- devtools/shell/jest/declarations.d.ts
@@ -1,1 +1,0 @@
-declare module "@gorhom/bottom-sheet/mock";
