--- devtools/shell/jest/mocks/react-native-worklets.js
@@ -1,1 +1,0 @@
-module.exports = require("react-native-worklets/lib/module/mock");
