+++ devtools/transport-panel/jest.config.js
@@ -1,32 +1,3 @@
-module.exports = {
-  testEnvironment: "jsdom",
-  roots: ["<rootDir>/src"],
-  transform: {
-    "^.+\\.(t|j)sx?$": [
-      "@swc/jest",
-      {
-        jsc: {
-          target: "esnext",
-          transform: {
-            react: {
-              runtime: "automatic",
-            },
-          },
-        },
-      },
-    ],
-  },
+module.exports = require("@support/jest-devtools-fixtures").createWebJestConfig({
   testTimeout: 30_000,
-  moduleFileExtensions: ["web.tsx", "web.ts", "tsx", "ts", "js", "jsx", "json", "node"],
-  testPathIgnorePatterns: ["\\.native\\.test\\."],
-  modulePaths: ["<rootDir>"],
-  setupFilesAfterEnv: ["<rootDir>/jest.setup.ts"],
-  transformIgnorePatterns: [
-    "node_modules/.pnpm/(?!(@ledgerhq\\+lumen-ui-react|@ledgerhq\\+lumen-design-core|@ledgerhq\\+lumen-utils-shared))",
-  ],
-  coverageReporters: ["json", ["lcov", { file: "lcov.info", projectRoot: "../../" }], "text"],
-  reporters: [
-    "default",
-    ["jest-sonar", { outputName: "sonar-executionTests-report.xml", reportedFilePath: "absolute" }],
-  ],
-};
+});
