--- devtools/transport-panel/jest/mocks/react-native-blur.tsx
@@ -1,4 +1,0 @@
-import React from "react";
-import { View } from "react-native";
-
-export const BlurView = ({ children }: { children?: React.ReactNode }) => <View>{children}</View>;
