+++ devtools/transport-panel/package.json
@@ -30,6 +30,7 @@
     "@devtools/transport": "workspace:*"
   },
   "devDependencies": {
+    "@support/jest-devtools-fixtures": "workspace:*",
     "@testing-library/react": "catalog:",
     "@testing-library/react-native": "catalog:",
     "react-test-renderer": "catalog:",
