+++ devtools/transport-panel/src/components/TransportPanelContent/TransportPanelContent.native.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { render, screen, fireEvent } from "jest/render";
+import { render, screen, fireEvent } from "jest/render.native";
 import { mockTransport, mockTransportPanelProps } from "jest/mocks/transport";
 import { TransportPanelContent } from "./TransportPanelContent";
